refactor: rename lib
