+++ apps/ac-mfe-profile/components/HeroSection/index.tsx
@@ -1,2 +1,2 @@
-import { HeroSection } from '@ac/ui-shared/lib/components/HeroSection';
+import { HeroSection } from '@ac/ui/modules/shared/components/HeroSection';
 export default HeroSection;
