+++ libs/ui/.babelrc
@@ -10,7 +10,5 @@
       }
     ]
   ],
-  "plugins": [
-    "@emotion/babel-plugin"
-  ]
+  "plugins": ["@emotion/babel-plugin"]
 }
