+++ libs/ui/.eslintrc.json
@@ -1,7 +1,7 @@
 {
   "extends": [
     "plugin:@nrwl/nx/react",
-    "../../../.eslintrc.json",
+    "../../.eslintrc.json",
     "plugin:cypress/recommended"
   ],
   "ignorePatterns": ["!**/*"],
