+++ libs/ui/README.md
@@ -1,0 +1,7 @@
+# ui
+
+This library was generated with [Nui](https://nui.dev).
+
+## Running unit tests
+
+Run `nui test ui` to euiecute the unit tests via [Jest](https://jestjs.io).
