+++ libs/ui/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@ac/ui",
+  "version": "0.0.1"
+}
