--- libs/ui/shared/README.md
@@ -1,7 +1,0 @@
-# ui-shared
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test ui-shared` to execute the unit tests via [Jest](https://jestjs.io).
