--- libs/ui/shared/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@ac/ui-shared",
-  "version": "0.0.1"
-}
