--- libs/ui/shared/src/index.ts
@@ -1,5 +1,0 @@
-import { HeroSection } from './lib/components/HeroSection';
-
-module.exports = {
-  HeroSection: HeroSection,
-};
