+++ libs/ui/src/index.ts
@@ -1,0 +1,5 @@
+import { HeroSection } from './lib/modules/shared/components/HeroSection';
+
+module.exports = {
+  HeroSection: HeroSection,
+};
