+++ libs/ui/src/lib/modules/shared/components/HeroSection/index.tsx
