+++ libs/ui/src/lib/modules/shared/components/HeroSection/mocks.ts
