+++ libs/ui/src/lib/modules/shared/components/HeroSection/types.ts
